`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: New `size_t` API.
Motivation for using `size_t` instead of `uint32_t`:

* It makes it easier to interoperate with C++ container APIs that
  typically use `size_t`.  In particular, it makes it easier to execute
  subsequent migration to `span` (`span::size` returns `size_t`).  See
  also WIP CL at https://crrev.com/c/5404364.
* It supports the `-Wshorten-64-to-32` work. See also crbug.com/1201109
  and
  https://docs.google.com/document/d/1CTbQ-5cQjnjU8aCOtLiA7G6P0i5C6HpSDNlSNq6nl5E/edit?usp=sharing

This CL introduces new `size_t`-based APIs in
`//mojo/public/c/system/data_pipe.h`.  Notes:

* To support incremental migration the old `uint32_t`-based APIs have
  been retained in `//mojo/public/cpp/system/data_pipe.h` for now.
  Follow-up CLs will migrate remaining users of the old APIs outside of
  `//mojo`.
* This CL does *not* change the C-level APIs because of the requirement
  to retain a stable ABI (see also the comments at the top of
  `//mojo/public/c/system/thunks.h`).
* This CL may have introduced new integer conversions - in those cases
  the CL attempts to have well-defined behavior (e.g. using
  `base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
  integer operations (e.g. old `int` => `uint32_t` / `size_t`
  conversions, or old additions that may potentially overflow).

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5414148
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284545}

+++ mojo/public/cpp/bindings/tests/handle_passing_unittest.cc
@@ -111,7 +111,7 @@ class SampleFactoryImpl : public sample::Factory {
     // Read the data from the pipe, writing the response (as a string) to
     // DidStuff2().
     ASSERT_TRUE(pipe.is_valid());
-    uint32_t data_size = 0;
+    size_t data_size = 0;
 
     MojoHandleSignalsState state;
     ASSERT_EQ(MOJO_RESULT_OK,
@@ -119,9 +119,9 @@ class SampleFactoryImpl : public sample::Factory {
     ASSERT_TRUE(state.satisfied_signals & MOJO_HANDLE_SIGNAL_READABLE);
     ASSERT_EQ(MOJO_RESULT_OK,
               pipe->ReadData(nullptr, &data_size, MOJO_READ_DATA_FLAG_QUERY));
-    ASSERT_NE(0, static_cast<int>(data_size));
+    ASSERT_NE(0u, data_size);
     char data[64];
-    ASSERT_LT(static_cast<int>(data_size), 64);
+    ASSERT_LT(data_size, 64u);
     ASSERT_EQ(MOJO_RESULT_OK, pipe->ReadData(data, &data_size,
                                              MOJO_READ_DATA_FLAG_ALL_OR_NONE));
 
@@ -268,7 +268,7 @@ TEST_P(HandlePassingTest, DataPipe) {
             CreateDataPipe(&options, producer_handle, consumer_handle));
   std::string expected_text_reply = "got it";
   // +1 for \0.
-  uint32_t data_size = static_cast<uint32_t>(expected_text_reply.size() + 1);
+  size_t data_size = expected_text_reply.size() + 1;
   ASSERT_EQ(MOJO_RESULT_OK,
             producer_handle->WriteData(expected_text_reply.c_str(), &data_size,
                                        MOJO_WRITE_DATA_FLAG_ALL_OR_NONE));
